Refactor pattern extracting to mir package
The notation package should not depend on mir
package, so the logic gor getting a Pattern
from a Score using PatternPosition belongs in mir.
